+++ kibot/out_dxf_sch_print.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2020-2023 Salvador E. Tropea
 # Copyright (c) 2020-2023 Instituto Nacional de Tecnología Industrial
-# Copyright (c) 2020 @nerdyscout
 # License: GPL-3.0
 # Project: KiBot (formerly KiPlot)
 """
@@ -13,36 +12,36 @@ Dependencies:
 """
 from .gs import GS
 from .out_any_sch_print import Any_SCH_PrintOptions
-from .misc import SVG_SCH_PRINT
+from .misc import DXF_SCH_PRINT
 from .macros import macros, document, output_class  # noqa: F401
 from . import log
 
 logger = log.get_logger()
 
 
-class SVG_SCH_PrintOptions(Any_SCH_PrintOptions):
+class DXF_SCH_PrintOptions(Any_SCH_PrintOptions):
     def __init__(self):
         with document:
             self.output = GS.def_global_output
-            """ Filename for the output SVG (%i=schematic, %x=svg) """
+            """ Filename for the output DXF (%i=schematic, %x=dxf)"""
         super().__init__()
-        self._expand_ext = 'svg'
-        self._exit_error = SVG_SCH_PRINT
+        self._expand_ext = 'dxf'
+        self._exit_error = DXF_SCH_PRINT
 
 
 @output_class
-class SVG_SCH_Print(BaseOutput):  # noqa: F821
-    """ SVG Schematic Print
-        Exports the schematic in a vectorized graphics format.
-        This is a format to document your schematic. """
+class DXF_SCH_Print(BaseOutput):  # noqa: F821
+    """ DXF Schematic Print (Drawing Exchange Format)
+        Exports the schematic to a format commonly used for CAD software.
+        This output is what you get from the 'File/Plot' menu in eeschema. """
     def __init__(self):
         super().__init__()
         with document:
-            self.options = SVG_SCH_PrintOptions
-            """ *[dict] Options for the `svg_sch_print` output """
+            self.options = DXF_SCH_PrintOptions
+            """ *[dict] Options for the `dxf_sch_print` output """
         self._sch_related = True
         self._category = 'Schematic/docs'
 
     @staticmethod
     def get_conf_examples(name, layers, templates):
-        return BaseOutput.simple_conf_examples(name, 'Schematic in SVG format', 'Schematic')  # noqa: F821
+        return BaseOutput.simple_conf_examples(name, 'Schematic in DXF format', 'Schematic')  # noqa: F821